bitcoin_pv_mining/requirements.txt
# Anforderungen für das PV-Mining Home Assistant Add-on
dash==2.6.2
plotly
requests
pyyaml
